requirements.txt
# Core Flask dependencies
Flask==2.3.3
Flask-Login==0.6.3

# Database
sqlite3

# Security
bcrypt==4.0.1

# AI API
groq==0.4.1

# Environment management
python-dotenv==1.0.0

# Development
Werkzeug==2.3.7
